Refactor rangy module: move _parse function to parse_old.py and update imports
