+++ rangy/__init__.py
@@ -2,9 +2,9 @@ from .const import ANY, AT_LEAST_ONE, EXACT, RANGE, COUNT_TYPES, INFINITY, ANY_C
 from .converters import Converter
 from .registry import ConverterRegistry
 from .builtins import register_builtins
-from .rangy import Rangy, _parse
+from .rangy import Rangy
 from .distribute import distribute
 from .parse import parse_range, _normalize_to_sequence, _convert_part
 
 register_builtins()
-__all__ = ["ANY", "AT_LEAST_ONE", "EXACT", "RANGE", "COUNT_TYPES", "INFINITY", "ANY_CHAR", "AT_LEAST_ONE_CHAR", "SPECIAL_CHARS", "Converter", "ConverterRegistry", "Rangy", "_parse", "distribute", "parse_range", "_normalize_to_sequence", "_convert_part"]
+__all__ = ["ANY", "AT_LEAST_ONE", "EXACT", "RANGE", "Rangy", "parse_range", "distribute", "Converter", "ConverterRegistry", "COUNT_TYPES", "INFINITY", "ANY_CHAR", "AT_LEAST_ONE_CHAR", "SPECIAL_CHARS", "_normalize_to_sequence", "_convert_part"]
